refactor: move logic from sync/index to sync/sync and use index for exporting
